# backend/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: EnglishLover API
  version: 1.0.0
  description: |
    EnglishLover 后端接口冻结契约。该契约是前端 `src/types/api.ts` 自动生成的唯一来源，字段、枚举、错误码和响应结构冻结后不得直接修改。
  x-contract-frozen: true
servers:
  - url: http://127.0.0.1:18080
    description: Stable local Compose test server
security:
  - accessCookie: []
tags:
  - name: Health
    description: 基础服务状态
  - name: Auth
    description: 用户认证与会话
  - name: Word
    description: 单词与学习队列
  - name: Review
    description: 学习复习与事件
  - name: Reading
    description: 阅读材料与跨模块队列
  - name: Penpal
    description: 受控笔友写信
  - name: Analytics
    description: 学习成果统计
paths:
  /api/v1/health:
    get:
      tags: [Health]
      operationId: getHealth
      x-api-id: API-HEALTH-GET
      x-requirement-ids: []
      x-test-ids: [TEST-OPENAPI-CONTRACT-01]
      x-nfr-ids: [NFR-STAB-02]
      summary: 获取服务健康状态
      security: []
      responses:
        '200':
          description: 服务可用
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
  /api/v1/auth/login:
    post:
      tags: [Auth]
      operationId: login
      x-api-id: API-AUTH-LOGIN
      x-requirement-ids: [REQ-AUTH-01]
      x-test-ids: [TEST-AUTH-LOGIN-01, TEST-OPENAPI-CONTRACT-01]
      x-nfr-ids: [NFR-SEC-04]
      summary: 登录并建立会话
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: 登录成功；令牌通过 HttpOnly Cookie 设置
          headers:
            Set-Cookie:
              description: access_token 与 refresh_token 采用 HttpOnly/Secure/SameSite Cookie
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/AuthInvalid'
  /api/v1/auth/refresh:
    post:
      tags: [Auth]
      operationId: refreshAuthToken
      x-api-id: API-AUTH-REFRESH
      x-requirement-ids: [REQ-AUTH-02]
      x-test-ids: [TEST-AUTH-SESSION-01, TEST-OPENAPI-CONTRACT-01]
      x-nfr-ids: [NFR-SEC-04]
      summary: 使用 Refresh Token 轮换会话
      security:
        - refreshCookie: []
      responses:
        '200':
          description: 刷新成功；返回当前用户摘要并重新设置 Cookie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/AuthRequired'
  /api/v1/auth/logout:
    post:
      tags: [Auth]
      operationId: logout
      x-api-id: API-AUTH-LOGOUT
      x-requirement-ids: [REQ-AUTH-03]
      x-test-ids: [TEST-AUTH-SESSION-01, TEST-OPENAPI-CONTRACT-01]
      x-nfr-ids: [NFR-SEC-04]
      summary: 退出登录并使 Refresh Token 失效
      responses:
        '204':
          description: 退出成功，无响应体
        '401':
          $ref: '#/components/responses/AuthRequired'
  /api/v1/auth/me:
    get:
      tags: [Auth]
      operationId: getCurrentUser
      x-api-id: API-AUTH-ME
      x-requirement-ids: [REQ-AUTH-04]
      x-test-ids: [TEST-AUTH-SESSION-01, TEST-OPENAPI-CONTRACT-01]
      x-nfr-ids: [NFR-SEC-01]
      summary: 获取当前登录用户
      responses:
        '200':
          description: 当前用户资料
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileResponse'
        '401':
          $ref: '#/components/responses/AuthRequired'
  /api/v1/words:
    get:
      tags: [Word]
      operationId: listWords
      x-api-id: API-WORD-LIST
      x-requirement-ids: [REQ-VOC-01]
      x-test-ids: [TEST-WORD-LIST-01, TEST-OPENAPI-CONTRACT-01]
      x-nfr-ids: [NFR-STAB-02]
      summary: 分页查询词库
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: stage
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/WordStage'
      responses:
        '200':
          description: 词库分页结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WordListResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/AuthRequired'
  /api/v1/reviews/submit:
    post:
      tags: [Review]
      operationId: submitReview
      x-api-id: API-REVIEW-SUBMIT
      x-requirement-ids: [REQ-VOC-03, REQ-VOC-04, REQ-VOC-05, REQ-VOC-08]
      x-test-ids: [TEST-WORD-REVIEW-01, TEST-WORD-RESUME-01, TEST-OPENAPI-CONTRACT-01]
      x-nfr-ids: [NFR-STAB-01]
      summary: 提交一次单词复习结果
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewSubmitRequest'
      responses:
        '200':
          description: 复习提交成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewSubmitResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/AuthRequired'
        '409':
          $ref: '#/components/responses/IdempotencyConflict'
  /api/v1/review-events:
    get:
      tags: [Review]
      operationId: listReviewEvents
      x-api-id: API-REVIEW-EVENTS
      x-requirement-ids: [REQ-VOC-05, REQ-REC-04]
      x-test-ids: [TEST-WORD-REVIEW-01, TEST-OPENAPI-CONTRACT-01]
      x-nfr-ids: [NFR-STAB-01]
      summary: 游标分页查询复习事件
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: 复习事件游标分页结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewEventListResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/AuthRequired'
  /api/v1/reading/articles:
    get:
      tags: [Reading]
      operationId: listReadingArticles
      x-api-id: API-READING-LIST
      x-requirement-ids: [REQ-RD-01, REQ-RD-02]
      x-test-ids: [TEST-READING-PRACTICE-01, TEST-OPENAPI-CONTRACT-01]
      x-nfr-ids: [NFR-STAB-02]
      summary: 分页查询阅读材料
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: 阅读材料分页结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadingArticleListResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/AuthRequired'
  /api/v1/reading/articles/{articleId}:
    get:
      tags: [Reading]
      operationId: getReadingArticle
      x-api-id: API-READING-DETAIL
      x-requirement-ids: [REQ-RD-03]
      x-test-ids: [TEST-READING-PRACTICE-01, TEST-OPENAPI-CONTRACT-01]
      x-nfr-ids: [NFR-STAB-02]
      summary: 获取阅读材料详情
      parameters:
        - $ref: '#/components/parameters/ArticleId'
      responses:
        '200':
          description: 阅读材料详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadingArticleResponse'
        '401':
          $ref: '#/components/responses/AuthRequired'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/reading/articles/{articleId}/words/{wordId}/queue:
    post:
      tags: [Reading, Word]
      operationId: addToWordLearningQueue
      x-api-id: API-READING-WORD-QUEUE
      x-requirement-ids: [REQ-RD-04, REQ-VOC-01]
      x-test-ids: [TEST-READING-WORD-QUEUE-01, TEST-OPENAPI-CONTRACT-01]
      x-nfr-ids: [NFR-STAB-01]
      summary: 从阅读模块将重点词加入单词待学习队列
      parameters:
        - $ref: '#/components/parameters/ArticleId'
        - $ref: '#/components/parameters/WordId'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: 加入成功或重复加入
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddToWordLearningQueueResponse'
        '401':
          $ref: '#/components/responses/AuthRequired'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/IdempotencyConflict'
  /api/v1/penpal/threads:
    get:
      tags: [Penpal]
      operationId: listPenpalThreads
      x-api-id: API-PENPAL-THREADS
      x-requirement-ids: [REQ-PAL-02, REQ-PAL-03, REQ-PAL-09]
      x-test-ids: [TEST-PENPAL-LETTER-01, TEST-PENPAL-RELATION-01, TEST-OPENAPI-CONTRACT-01]
      x-nfr-ids: [NFR-SEC-02]
      summary: 分页查询受控笔友会话
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: 笔友会话分页结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PenpalThreadListResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/AuthRequired'
  /api/v1/penpal/letters:
    post:
      tags: [Penpal]
      operationId: sendPenpalLetter
      x-api-id: API-PENPAL-LETTER-SEND
      x-requirement-ids: [REQ-PAL-04, REQ-PAL-06, REQ-PAL-08, REQ-REC-03]
      x-test-ids: [TEST-PENPAL-LETTER-01, TEST-PENPAL-SAFETY-01, TEST-OPENAPI-CONTRACT-01]
      x-nfr-ids: [NFR-SEC-03]
      summary: 发送受控笔友信件
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendPenpalLetterRequest'
      responses:
        '200':
          description: 信件发送成功并触发 penpal_activity_completed 事件
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SendPenpalLetterResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/AuthRequired'
        '409':
          $ref: '#/components/responses/IdempotencyConflict'
  /api/v1/analytics/daily-summary:
    get:
      tags: [Analytics]
      operationId: getDailySummary
      x-api-id: API-ANALYTICS-DAILY-SUMMARY
      x-requirement-ids: [REQ-REC-01, REQ-REC-02, REQ-REC-03, REQ-NAV-02]
      x-test-ids: [TEST-ANALYTICS-SUMMARY-01, TEST-OPENAPI-CONTRACT-01]
      x-nfr-ids: [NFR-STAB-01]
      summary: 查询指定日期的学习成果摘要
      parameters:
        - name: date
          in: query
          required: true
          description: ISO 8601 日期，格式为 YYYY-MM-DD
          schema:
            type: string
            format: date
      responses:
        '200':
          description: 当日学习成果摘要
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DailySummaryResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/AuthRequired'
components:
  securitySchemes:
    accessCookie:
      type: apiKey
      in: cookie
      name: access_token
      description: 短有效期 Access Token，建议 15 分钟。
    refreshCookie:
      type: apiKey
      in: cookie
      name: refresh_token
      description: 长有效期 Refresh Token，建议 7-30 天并启用轮换。
  parameters:
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Cursor:
      name: cursor
      in: query
      required: false
      schema:
        type: string
        minLength: 1
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: 高风险写接口必须携带的幂等键。
      schema:
        type: string
        minLength: 8
        maxLength: 128
    ArticleId:
      name: articleId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    WordId:
      name: wordId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    AuthRequired:
      description: 未登录或会话缺失
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    AuthInvalid:
      description: 认证凭据无效
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    ValidationError:
      description: 请求参数不合法
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    IdempotencyConflict:
      description: 幂等键冲突
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
  schemas:
    ErrorCode:
      type: string
      enum:
        - AUTH_REQUIRED
        - AUTH_INVALID
        - FORBIDDEN
        - VALIDATION_ERROR
        - NOT_FOUND
        - CONFLICT
        - IDEMPOTENCY_CONFLICT
        - RATE_LIMITED
        - INTERNAL_ERROR
    LearningModule:
      type: string
      enum: [word, reading, writing]
    SessionStatus:
      type: string
      enum: [active, completed, interrupted, discarded]
    WordStage:
      type: string
      enum: [cet4, cet6, kaoyan, toefl, ielts, general]
    ReviewMode:
      type: string
      enum: [recognition, recall, spelling]
    ReviewRating:
      type: integer
      minimum: 1
      maximum: 5
    WordLearningSource:
      type: string
      enum: [reading, manual, system]
    WordQueueStatus:
      type: string
      enum: [success, duplicate]
    PenpalActivityType:
      type: string
      enum: [letter_sent]
    PageMeta:
      type: object
      required: [page, pageSize, total]
      properties:
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
    CursorMeta:
      type: object
      required: [limit]
      properties:
        cursor:
          type: string
        limit:
          type: integer
          minimum: 1
        nextCursor:
          type: string
          nullable: true
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
        details:
          type: object
          additionalProperties: true
    ErrorEnvelope:
      type: object
      required: [requestId, error]
      properties:
        requestId:
          type: string
        error:
          $ref: '#/components/schemas/ApiError'
    HealthData:
      type: object
      required: [status, service, version]
      properties:
        status:
          type: string
          enum: [ok]
        service:
          type: string
          enum: [englishlover-api]
        version:
          type: string
    HealthResponse:
      type: object
      required: [requestId, data]
      properties:
        requestId:
          type: string
        data:
          $ref: '#/components/schemas/HealthData'
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
    UserRole:
      type: string
      enum: [learner, content_admin, admin, ops]
    UserProfile:
      type: object
      required: [userId, email, displayName, role]
      properties:
        userId:
          type: string
          format: uuid
        email:
          type: string
          format: email
        displayName:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
    AuthResponse:
      type: object
      required: [requestId, data]
      properties:
        requestId:
          type: string
        data:
          type: object
          required: [user]
          properties:
            user:
              $ref: '#/components/schemas/UserProfile'
    UserProfileResponse:
      type: object
      required: [requestId, data]
      properties:
        requestId:
          type: string
        data:
          $ref: '#/components/schemas/UserProfile'
    WordSummary:
      type: object
      required: [wordId, text, stage]
      properties:
        wordId:
          type: string
          format: uuid
        text:
          type: string
        phonetic:
          type: string
        stage:
          $ref: '#/components/schemas/WordStage'
    WordListData:
      type: object
      required: [items, page, pageSize, total]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/WordSummary'
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
    WordListResponse:
      type: object
      required: [requestId, data]
      properties:
        requestId:
          type: string
        data:
          $ref: '#/components/schemas/WordListData'
    ReviewSubmitRequest:
      type: object
      required: [wordId, mode, rating, isCorrect, durationMs, clientOccurredAt]
      properties:
        wordId:
          type: string
          format: uuid
        mode:
          $ref: '#/components/schemas/ReviewMode'
        rating:
          $ref: '#/components/schemas/ReviewRating'
        isCorrect:
          type: boolean
        durationMs:
          type: integer
          minimum: 0
        clientOccurredAt:
          type: string
          format: date-time
    ReviewSubmitResult:
      type: object
      required: [reviewEventId, status]
      properties:
        reviewEventId:
          type: string
          format: uuid
        status:
          type: string
          enum: [accepted]
        nextReviewAt:
          type: string
          format: date-time
    ReviewSubmitResponse:
      type: object
      required: [requestId, data]
      properties:
        requestId:
          type: string
        data:
          $ref: '#/components/schemas/ReviewSubmitResult'
    ReviewEvent:
      type: object
      required: [eventId, module, occurredAt]
      properties:
        eventId:
          type: string
          format: uuid
        module:
          $ref: '#/components/schemas/LearningModule'
        occurredAt:
          type: string
          format: date-time
    ReviewEventListData:
      type: object
      required: [items, limit]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ReviewEvent'
        cursor:
          type: string
        limit:
          type: integer
          minimum: 1
        nextCursor:
          type: string
          nullable: true
    ReviewEventListResponse:
      type: object
      required: [requestId, data]
      properties:
        requestId:
          type: string
        data:
          $ref: '#/components/schemas/ReviewEventListData'
    ReadingArticleSummary:
      type: object
      required: [articleId, title]
      properties:
        articleId:
          type: string
          format: uuid
        title:
          type: string
        level:
          type: string
    ReadingArticleListData:
      type: object
      required: [items, page, pageSize, total]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ReadingArticleSummary'
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
    ReadingArticleListResponse:
      type: object
      required: [requestId, data]
      properties:
        requestId:
          type: string
        data:
          $ref: '#/components/schemas/ReadingArticleListData'
    ReadingArticleDetail:
      type: object
      required: [articleId, title, content]
      properties:
        articleId:
          type: string
          format: uuid
        title:
          type: string
        content:
          type: string
    ReadingArticleResponse:
      type: object
      required: [requestId, data]
      properties:
        requestId:
          type: string
        data:
          $ref: '#/components/schemas/ReadingArticleDetail'
    AddToWordLearningQueueResult:
      type: object
      required: [status]
      properties:
        status:
          $ref: '#/components/schemas/WordQueueStatus'
        source:
          $ref: '#/components/schemas/WordLearningSource'
    AddToWordLearningQueueResponse:
      type: object
      required: [requestId, data]
      properties:
        requestId:
          type: string
        data:
          $ref: '#/components/schemas/AddToWordLearningQueueResult'
    PenpalThreadSummary:
      type: object
      required: [threadId, status]
      properties:
        threadId:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/SessionStatus'
    PenpalThreadListData:
      type: object
      required: [items, page, pageSize, total]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PenpalThreadSummary'
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
    PenpalThreadListResponse:
      type: object
      required: [requestId, data]
      properties:
        requestId:
          type: string
        data:
          $ref: '#/components/schemas/PenpalThreadListData'
    SendPenpalLetterRequest:
      type: object
      required: [threadId, body]
      properties:
        threadId:
          type: string
          format: uuid
        body:
          type: string
          minLength: 1
          maxLength: 2000
    SendPenpalLetterResult:
      type: object
      required: [letterId, activityType]
      properties:
        letterId:
          type: string
          format: uuid
        activityType:
          $ref: '#/components/schemas/PenpalActivityType'
    SendPenpalLetterResponse:
      type: object
      required: [requestId, data]
      properties:
        requestId:
          type: string
        data:
          $ref: '#/components/schemas/SendPenpalLetterResult'
    DailySummary:
      type: object
      required: [date, wordCompletedCount, readingCompletedCount, writingCompletedCount, taskCompletionRate, streakIncluded]
      properties:
        date:
          type: string
          format: date
        wordCompletedCount:
          type: integer
          minimum: 0
        readingCompletedCount:
          type: integer
          minimum: 0
        writingCompletedCount:
          type: integer
          minimum: 0
        taskCompletionRate:
          type: number
          minimum: 0
          maximum: 1
        streakIncluded:
          type: boolean
    DailySummaryResponse:
      type: object
      required: [requestId, data]
      properties:
        requestId:
          type: string
        data:
          $ref: '#/components/schemas/DailySummary'